+++ src/libraries/MathLib.sol
@@ -3,19 +3,19 @@ pragma solidity ^0.8.0;
 
 /// @notice Maths utils.
 library MathLib {
-    uint internal constant WAD = 1e18;
+    uint256 internal constant WAD = 1e18;
 
-    function min(uint x, uint y) internal pure returns (uint z) {
+    function min(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = x < y ? x : y;
     }
 
     /// @dev Rounds towards zero.
-    function wMul(uint x, uint y) internal pure returns (uint z) {
+    function wMul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = (x * y) / WAD;
     }
 
     /// @dev Rounds towards zero.
-    function wDiv(uint x, uint y) internal pure returns (uint z) {
+    function wDiv(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = (x * WAD) / y;
     }
 }
